Update Main.java
This is a menu (with minimal changes) that should work with the other parts of the program in order to send information to the API we are using. I'm not quite sure what the API returns so I have not done any formatting for the returned information.

Hopefully the comments I left in the program explain enough about what is happening, if not I can be more detailed.

P.S. Hopefully I did things correctly, I was looking at GitHub's Documentation to try to figure this out.